Wrap SVG into an <img> to make it selectable
This also removes the need to mangle the ids inside each SVG.

Unfortunately this takes away the possibility to style the
formulas with CSS, and we need that to make sure night mode
works. The solution is to invert the image with CSS filters.
Note that this will mess up all the colors aside from black.
Let's revisit this if we'll need colored LaTeX.