Speed up tensor conversion with many integral families
It seems that `id BID=1` is *much* slower than `id BID^x?=1` if the
exponent is large, because FORM apparently expands the exponent
into multiple variable copies, which means each term needs to
be copied around many times. A good chunk of the reduction time
can be spent on this.