Make QGraf return the correct diagram signs
QGraf only returns the correct signs of the diagrams if fields in
the vertices follow a specific order: <anti-fermion> <fermion> <X>.
For example one should use "[Q, q, g]" instead of "[g, q, Q]" in
the QGraf model file.

Before commit 34b231d we disregarded the diagram signs that
QGraf computes and counted fermion loops instead. After this
logic was removed, we need to also swap the order of fields in
the vertices so that the signs QGraf computes match the previous
(correct) values.

The Feynman rules themselves are intact, with just p1<->p3
exchanged when "gqQ" is rearranged as "Qqg".