Add option to use additional filters for QGraf in Diagrams[...]
Add a new option QGrafExtra to Diagrams[] to allow usage of additional
filters, e.g. employing v_functions, during diagram generation with
QGraf. This is explained in detail in sections 18 and 19 of the QGRAF
4.0 manual.

A new global variable $QGrafFilter is the default value of this option.
All model files should set it (even to an empty value), to override
whatever the previously loaded model file might have used.

Co-Authored-By: Augustin Vestner <augustin.vestner@kit.edu>